Gradient descent initial

+++ include/machine_learning_algos/LinearRegression.h
@@ -2,9 +2,10 @@
 
 class LinearRegression {
     public:
-        LinearRegression(double learning_rate = 0.01, int threshold = 0.00001) : learning_rate(learning_rate), threshold(threshold) {};
+        LinearRegression(double learning_rate = 0.01, double threshold = 0.00001) : learning_rate(learning_rate), threshold(threshold) {};
 
         void fit(const std::vector<std::vector<double>>& X, const std::vector<int>& y);
+
         std::vector<int> predict(const std::vector<std::vector<double>>& X);
     private:
         enum WeightType {
@@ -19,7 +20,7 @@ class LinearRegression {
         std::vector<Weight> weights;
 
         double learning_rate;
-        int threshold;
+        double threshold;
 
         double MSE_loss(const std::vector<int>& y_true, const std::vector<int>& y_pred);
